requirements.txt
# Core dependencies for AI Context Craft
# Core dependencies
pyyaml>=6.0
pathspec>=0.11.0

# For comment stripping
tree-sitter>=0.20.0
tree-sitter-languages>=1.5.0